implemented api authorization

+++ backend/controllers/noteController.js
@@ -1,6 +1,7 @@
 const asyncHandler = require("express-async-handler");
 const mongoose = require("mongoose");
 const Note = require("../models/noteModel");
+const User = require("../models/userModel");
 
 // @desc GET NOTES
 // @route GET /api/notes
@@ -9,7 +10,8 @@ const { error } = require("console");
 const noteModel = require("../models/noteModel");
 
 const getNotes = asyncHandler(async (req, res) => {
-  const notes = await Note.find();
+  const notes = await Note.find({ user: req.user.id });
+
   res.status(200).json(notes);
 });
 
@@ -24,6 +26,7 @@ const setNote = asyncHandler(async (req, res) => {
 
   const note = await Note.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(note);
 });
@@ -39,6 +42,20 @@ const updateNote = asyncHandler(async (req, res) => {
     throw new Error("Note Not Found");
   }
 
+  const user = await User.findById(req.user.id);
+
+  // Check is user exists
+  if (!user) {
+    res.status(401);
+    throw new Error("User not authorized to update");
+  }
+
+  // Check if logged in user matches the note user
+  if (note.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not AUTHORIZED");
+  }
+
   const updatedNote = await Note.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -50,15 +67,28 @@ const updateNote = asyncHandler(async (req, res) => {
 // @route DELETE /api/notes/:id
 // @access Private
 const deleteNote = asyncHandler(async (req, res) => {
-  // const note = await Note.findById(req.params.id);
-  const note = await Note.findByIdAndDelete(req.params.id);
+  const note = await Note.findById(req.params.id);
 
   if (!note) {
     res.status(400);
     throw new Error("Note Not Found");
   }
+  const user = await User.findById(req.user.id);
+
+  // Check is user exists
+  if (!user) {
+    res.status(401);
+    throw new Error("User not authorized to delete");
+  }
+
+  // Check if logged in user matches the note user
+  if (note.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not AUTHORIZED");
+  }
 
   // await note.remove();
+  await Note.findByIdAndDelete(req.params.id);
 
   res
     .status(200)
